+++ common/visibility/validate_search_attribute_key_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package validator
+package visibility
 
 import (
 	"testing"
